+++ packages/editor/src/editor/PortableTextEditor.tsx
@@ -2,15 +2,13 @@ import type {
   PortableTextBlock,
   PortableTextChild,
   PortableTextObject,
+  Schema,
 } from '@portabletext/schema'
-import {Subject} from 'rxjs'
 import type {
   AddedAnnotationPaths,
   EditableAPI,
   EditableAPIDeleteOptions,
-  EditorChanges,
   EditorSelection,
-  PortableTextMemberSchemaTypes,
 } from '../types/editor'
 import type {Path} from '../types/paths'
 import type {InternalEditor} from './create-editor'
@@ -34,13 +32,9 @@ import type {InternalEditor} from './create-editor'
  */
 export class PortableTextEditor {
   /**
-   * An observable of all the editor changes.
-   */
-  public change$: EditorChanges = new Subject()
-  /**
    * A lookup table for all the relevant schema types for this portable text type.
    */
-  public schemaTypes: PortableTextMemberSchemaTypes
+  public schemaTypes: Schema
   /**
    * The editor instance
    */
@@ -52,9 +46,8 @@ export class PortableTextEditor {
 
   constructor(config: {editor: InternalEditor}) {
     this.editor = config.editor
-    this.schemaTypes = config.editor._internal.editorActor
-      .getSnapshot()
-      .context.getLegacySchema()
+    this.schemaTypes =
+      config.editor._internal.editorActor.getSnapshot().context.schema
     this.editable = config.editor._internal.editable
   }