+++ packages/editor/tests/withEditableAPIDelete.test.tsx
@@ -1,7 +1,7 @@
 import {createRef, type RefObject} from 'react'
 import {describe, expect, it, vi} from 'vitest'
 import {PortableTextEditor} from '../src/editor/PortableTextEditor'
-import {InternalChange$Plugin} from '../src/plugins/plugin.internal.change-ref'
+import {EventListenerPlugin} from '../src/plugins'
 import {InternalPortableTextEditorRefPlugin} from '../src/plugins/plugin.internal.portable-text-editor-ref'
 import {createTestEditor} from '../src/test/vitest'
 
@@ -49,7 +49,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -59,7 +59,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await vi.waitFor(() => {
       if (editorRef.current) {
         expect(onChange).toHaveBeenCalledWith({
-          type: 'value',
+          type: 'value changed',
           value: initialValue,
         })
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
@@ -105,7 +105,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -114,7 +114,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
 
     await vi.waitFor(() => {
       expect(onChange).toHaveBeenCalledWith({
-        type: 'value',
+        type: 'value changed',
         value: initialValue,
       })
       expect(onChange).toHaveBeenCalledWith({type: 'ready'})
@@ -162,7 +162,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -172,7 +172,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await vi.waitFor(() => {
       if (editorRef.current) {
         expect(onChange).toHaveBeenCalledWith({
-          type: 'value',
+          type: 'value changed',
           value: initialValue,
         })
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
@@ -241,7 +241,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -251,7 +251,7 @@ describe('plugin:withEditableAPI: .delete()', () => {
     await vi.waitFor(() => {
       if (editorRef.current) {
         expect(onChange).toHaveBeenCalledWith({
-          type: 'value',
+          type: 'value changed',
           value: initialValue,
         })
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})