+++ packages/editor/tests/withEditableAPIInsert.test.tsx
@@ -2,7 +2,7 @@ import {defineSchema} from '@portabletext/schema'
 import {createRef, type RefObject} from 'react'
 import {describe, expect, it, vi} from 'vitest'
 import {PortableTextEditor} from '../src/editor/PortableTextEditor'
-import {InternalChange$Plugin} from '../src/plugins/plugin.internal.change-ref'
+import {EventListenerPlugin} from '../src/plugins'
 import {InternalPortableTextEditorRefPlugin} from '../src/plugins/plugin.internal.portable-text-editor-ref'
 import {createTestEditor} from '../src/test/vitest'
 
@@ -62,7 +62,7 @@ describe('plugin:withEditableAPI: .insertChild()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -77,7 +77,7 @@ describe('plugin:withEditableAPI: .insertChild()', () => {
     await vi.waitFor(() => {
       if (editorRef.current) {
         expect(onChange).toHaveBeenCalledWith({
-          type: 'value',
+          type: 'value changed',
           value: initialValue,
         })
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
@@ -210,7 +210,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -224,7 +224,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
 
     await vi.waitFor(() => {
       if (editorRef.current) {
-        expect(onChange).toHaveBeenCalledWith({type: 'value', value})
+        expect(onChange).toHaveBeenCalledWith({type: 'value changed', value})
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
       }
     })
@@ -268,7 +268,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -283,7 +283,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await vi.waitFor(() => {
       if (editorRef.current) {
         expect(onChange).toHaveBeenCalledWith({
-          type: 'value',
+          type: 'value changed',
           value: initialValue,
         })
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
@@ -330,7 +330,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -345,7 +345,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await vi.waitFor(() => {
       if (editorRef.current) {
         expect(onChange).toHaveBeenCalledWith({
-          type: 'value',
+          type: 'value changed',
           value: initialValue,
         })
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
@@ -399,7 +399,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -413,7 +413,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
 
     await vi.waitFor(() => {
       if (editorRef.current) {
-        expect(onChange).toHaveBeenCalledWith({type: 'value', value})
+        expect(onChange).toHaveBeenCalledWith({type: 'value changed', value})
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
       }
     })
@@ -466,7 +466,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -480,7 +480,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
 
     await vi.waitFor(() => {
       if (editorRef.current) {
-        expect(onChange).toHaveBeenCalledWith({type: 'value', value})
+        expect(onChange).toHaveBeenCalledWith({type: 'value changed', value})
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
       }
     })
@@ -528,7 +528,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
     await createTestEditor({
       children: (
         <>
-          <InternalChange$Plugin onChange={onChange} />
+          <EventListenerPlugin on={onChange} />
           <InternalPortableTextEditorRefPlugin ref={editorRef} />
         </>
       ),
@@ -542,7 +542,7 @@ describe('plugin:withEditableAPI: .insertBlock()', () => {
 
     await vi.waitFor(() => {
       if (editorRef.current) {
-        expect(onChange).toHaveBeenCalledWith({type: 'value', value})
+        expect(onChange).toHaveBeenCalledWith({type: 'value changed', value})
         expect(onChange).toHaveBeenCalledWith({type: 'ready'})
       }
     })